requirements.txt
torch>=2.2.0
torchvision>=0.17.0
diffusers>=0.30.0
transformers>=4.42.0,<4.50.0
accelerate>=0.31.0
safetensors>=0.4.3
huggingface_hub>=0.23.0
numpy>=1.24.0
scipy>=1.10.0
Pillow>=10.0.0
tqdm>=4.66.0
pyyaml>=6.0.1
pandas>=2.0.0
einops>=0.8.0
lpips>=0.1.4
piq>=0.8.0
requests>=2.31.0
imageio>=2.34.0
